Updated nifi deploy script.

+++ bin/nifi.fast.deploy.sh
@@ -1,20 +1,4 @@
 #!/bin/bash
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
 set -e
 
 SKIP_BUILD=false
@@ -34,6 +18,14 @@ do
             ;;
         -v)
             set -x
+            ;;
+        -h)
+            echo "Options: --skip (-s) to skip building and only deploy."
+            exit 0
+            ;;
+        --help)
+            echo "Options: --skip (-s) to skip building and only deploy."
+            exit 0
             ;;
         *)
             echo "Unknown option passed: $key"
@@ -57,7 +49,7 @@ nifi_name=`readlink -f /opt/nifi | xargs basename`
 
 build_dir=`ls -1 nifi-assembly/target/ | grep -e nifi.*bin | grep -v zip | xargs basename`
 build_name=`ls -1 nifi-assembly/target/$build_dir | xargs basename`
-if [ -d "/opt/$nifi_name" ]; then
+if [ -n "$nifi_name" ] && [ -d "/opt/$nifi_name" ]; then
     rm -f /opt/nifi/conf/archive/*
     /opt/nifi/bin/nifi.sh stop
